Fix for print_number_str
print_number_str would cause digit 'artifacts' if used to display some
number of digits, then used to display less digits in a following cycle.
The 'extra' digits would not be 'cleared'.

This fix pads the input string with spaces (based on justify_right
parameter) in order to push a blank digit to the buffer. This way, all 4
digit positions will be affected and there will be no artifacts.